+++ CHANGELOG.md
@@ -8,8 +8,11 @@
 
 - Make VocabularyTermsWidget orderable @ksuess
 - Get widget by tagged values @ksuess
+- Added custom option to SelectWidget to render custom optionss (for example with icons) @giuliaghisini
 
 ### Bugfix
+
+- Fixed selected widget to use isMulti prop @giuliaghisini
 
 ### Internal